이전 과정/requirements.txt
aiohttp==3.7.4.post0
anyio==3.3.0
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.2.1
backcall==0.2.0
bleach==4.0.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
colorama==0.4.4
contextvars==2.4
cryptography==3.4.7

decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
hyperlink==21.0.0
idna==3.2
idna-ssl==1.1.0
immutables==0.16
importlib-metadata==4.6.3
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyter-server-proxy==3.1.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.1
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.2
numpy==1.18.1
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pywin32==301
pywinpty==1.1.3
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.9.0
requests==2.26.0
requests-unixsocket==0.2.0
Send2Trash==1.7.1
simpervisor==0.4
six==1.16.0
sniffio==1.2.0
terminado==0.10.1
testpath==0.5.0

torchaudio==0.9.0

tornado==6.1
traitlets==4.3.3
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
vpython==7.6.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.1
widgetsnbextension==3.5.1
yarl==1.6.3
zipp==3.5.0
dataclasses==0.8
torch==1.9.0+cu102
torchvision==0.10.0+cu102